split segmener into two files
